refactor(theta): colocate Jaccard result types
Theta and Tuple Jaccard operators returned a type that callers had to import from the implementation-oriented `thetacommon` module. That leaked the internal sharing arrangement and made each sketch family appear incomplete.

Re-export the same `JaccardSimilarity` type from both `theta` and `tuple`, and make the common module crate-private. Only the import path changes; the result representation and operator behavior remain identical.
